# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Game Save API
  version: 1.0.0
  description: API pour la gestion des utilisateurs, jeux, associations et sauvegardes.
servers:
  - url: http://localhost:8000
    description: Local server
paths:
...
  /addGame/{title}/{steam_id}:
    post:
      summary: Ajoute un nouveau jeu
      operationId: addGame
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
          description: Titre du jeu
        - name: steam_id
          in: path
          required: true
          schema:
            type: string
          description: ID Steam du jeu
      responses:
        '200':
          description: Jeu ajouté avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  game_id:
                    type: integer
                  title:
                    type: string
                  steam_id:
                    type: string
        '400':
          description: Le jeu existe déjà.
...
  /games/{game_id}:
    get:
      summary: Récupère un jeu par ID
      operationId: getGameById
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: integer
          description: ID du jeu
      responses:
        '200':
          description: Détails du jeu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Jeu non trouvé.
...
  /saves/{save_id}:
    get:
      summary: Récupère une sauvegarde par ID
      operationId: getSaveById
      parameters:
        - name: save_id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la sauvegarde
      responses:
        '200':
          description: Détails de la sauvegarde
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Save'
        '404':
          description: Sauvegarde non trouvée.
components:
  schemas:
...
    Game:
      type: object
      properties:
        game_id:
          type: integer
        title:
          type: string
        steam_id:
          type: string
...
    Save:
      type: object
      properties:
        save_id:
          type: integer
        user_game_id:
          type: integer
        save_path:
          type: string
        save_date:
          type: string
          format: date-time